#8 Remove multiple now unnecessary sample application and tests.
The library matured enough to be able to connect the dots and get rid of many intermediate tests and classes that did not fully represent how the library can be used from the client's perspective.

From now on it should be possible to write tests from the end user perspective in a more black box API-based fashion.