#8 Bring back the previous functionality of the demo app (informing about data sent) that was suppressed by introduction of the delineation.
Added a calculated window size to the confirmation message as it can be useful to apply flow control inside the app.

DataSent event, such as
DataSent{port=9999, connectionId=0, bytesSent=25, totalBytesSent=301, totalBytesBuffered=301, commandId=-1, sendBufferSize=2626560, windowSizeInBytes=2626560}

is a confirmation message from the transport layer that the message
was attempted to send and is currently in on of the buffers between
the sender and the receiver. windowSizeInBytes=2626560 informs, that
at the moment of sending this confirmation, up to 2626560 bytes
can be sent without risking a back pressure, even if the receiving
party stopped consuming the data whatsoever. After the receiving
party stopped consuming the data and the network buffers sere filled,
this number should start shrinking, at which point the back pressure
can be applied to the upstream service if necessary to avoid data loss.
The window size is updated from the Transport layer on the same box/process,
so this the app should be notified promptly, early enough to act on
this information to prevent negative consequences of filling up the window.