#11 Do not send command failure when exception occurred.
In theory it was a nice idea to treat an unexpected failure as yet another message. However, this introduces a risk of infinite cycle when the handling application in reaction to the failure retries the command or tries to recover. This can lead to an infinite cycle and result in a stack overflow. Also, it obfuscates the place when the exception happened.

It is still a good idea to send such failure as a message but only when the handler does not share the stack with the code that threw an exception (ring buffer message driven application can be such a place)