#11 Do not work automatically right after handling a command
This is another example of a nice idea that can have unintended consequences when the work part involves handling another command - causing a loop.

This fix increases safety at the cost of a slight inconvenience of calling work after handling a command if necessary.

Work should be called quite regularly,
so this only makes a difference in unit tests.