#6 Extra logging for RoundTripTimeTest
Warning: The numbers below are not affected enough by GC pauses and overall jitter, as both sender and the receiver are run
on the same JVM, which causes a coordinated omission problem. When the receiver is not ready to receive more messages due to
a pause, at about the same time the producer conveniently do not send them due to the same pause. Two separate processes running
on two separate CPUs are necessary to present numbers closer to the reality.

Current numbers for the loopback RoundTripTimeTest are as follows.

In-stack transport integration:

419,278 messages reads + writes per second on aa single CPU
Exchanged 20000000 messages at a rate of 209,639 messages per second  which took 95 seconds
99.99th percentile is 51 microseconds

RingBuffer transport integration + serialization

376,136  messages reads + writes per second on a single CPU
Exchanged 20000000 messages at a rate of 188,068 messages per second  which took 106 seconds
99.99th percentile is 88 microseconds