#8 Support all defined delineation types when sending and receiving messages.
Turning messages into data stream and back
is not a user-provided transport application concern. It is calculated automatically when using SendMessage command and MessageReceived event.

No longer enforce that the fixed length message has correct size as the sender may
want to include more than one message in a batch.