#6 Verify the deterministic nature of the application
This library is ready to be used by event source applications and this test shows an example of how to capture the log. It is not a production grade way of capturing it and is not persisted, but one can get the idea.

The verification is done by capturing all events and commands in the log, then reading all the captured events by the newly created application and verify that the new log is the same as the original one.