#18 Detect Tcp Gateway by streamId
By defining a streamId for Tcp Gateway, clustered service
can easily create a session for this client and ignore non related ones.

By allowing more than one Tcp Gateway streamId, clustered service can decide to distribute load across multiple gateways, should such need arises. However, as this is done via a static configuration, it does not complicate logic if one stream is sufficient and a single Tcp Gateway session lives as long as the node is alive (and as long as the Tcp Gateway is around)