#18 Use TransportApplicationTest, but this time with the application inside Aeron cluster.
When you compare both tests, they verify the same simple behaviour that binds to a port, accepts connections and echos back messages sent by a tcp client.

The difference is that now one of the tests puts the transport application inside the cluster (keeping the same API). By putting a transport application inside the cluster one can leverage the properties of the cluster, such as replayability of the messages, audit log or high availability at the cost of more hardware and storage utilisation.