#18 Use transport application type that does not require constant waking up if not necessary
When transport application is on duty responsible for e.g. polling for new things to do (checking if there are new events or new data on the socket), it needs an appropriate interface (TransportApplicationOnDuty). When it is always notified and does not need to be proactive (e.g. when it is created inside a framework that deals with such concerns, e.g. Aeron Cluster), it is not on duty and has an appropriate interface (TransportApplication)