#22 Finish the first cut of the fake market
All important messages are now in place. The PnL module can be moved
to market participants as they now receive enough data to calculate it themselves. The market currently moves no faster than passing time, ie.
for millisecond resolution it is once per millisecond at most. This needs
to be changed to apply more realistic load of the market data (but avoid busy spinning sending price updates as it would render the market useless when other messages can't be processed)