Configure offsets of the metadata fields.
This allows measuring latency distribution of any TCP workflow by injecting the metadata into existing payloads of any kind,
as long as 3 places (with a minimum length of 8 bytes, 8 bytes and 4 bytes respectively) can be found and their meaning can be ignored or does not break other protocol layers.

One endpoint needs to instantiate a Probe and another endpoint needs
to instantiate a ProbeClient.

The simplest example is SourcingConnectionApplication and EchoApplication respectively, when probing is its only purpose.